hug_api/requirements.txt
hug
mysql-connector-python
hug_middleware_cors
